requirements-cpu.txt
zarr
fsspec[gcs]
jax[cpu]==0.4.26
einops
hydra-core
clearml
clearml-agent
typeguard==4.1.5
dm-haiku
ftfy
transformers  # For Huggingface data loader
datasets      # For Huggingface data loader
torch[cpu]    # For Huggingface data loader